feat(tailwind): migrated icons + icons buttons + prettier tailwind :rocket:
