+++ src/components/design-system/atoms/icons/down-arrow-icon.tsx
@@ -1,25 +1,11 @@
-"use client";
-
 import { BiChevronDown } from "react-icons/bi";
-import styled from "styled-components";
-import { mediaQuery } from "../../utils/media-query";
-
-const ArrowIcon = styled.div`
-  width: 30px;
-  height: 30px;
-  color: ${(props) => props.theme.colors.textAbovePrimaryColor};
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  ${mediaQuery.minWidth.md} {
-    width: 50px;
-    height: 50px;
-  }
-`;
+import { RoundedIcon } from "./icon";
 
 export const DownArrowIcon = () => (
-    <ArrowIcon>
-      <BiChevronDown className='size-7 md:size-9' title="Scroll to next section" />
-    </ArrowIcon>
+  <RoundedIcon className="text-text-above-primary flex animate-bounce items-center justify-center hover:shadow-lg">
+    <BiChevronDown
+      className="size-7 md:size-9"
+      title="Scroll to next section"
+    />
+  </RoundedIcon>
 );
