+++ src/components/design-system/molecules/buttons/chat-button.tsx
@@ -2,28 +2,8 @@
 
 import { slugs } from "@/types/slug";
 import Link from "next/link";
-import styled from "styled-components";
 import { usePathname } from "next/navigation";
 import { ChatIcon } from "../../atoms/icons/chat-icon";
-import { mediaQuery } from "../../utils/media-query";
-
-export const ChatButton = () => (
-  <Link target="_blank" href={slugs.chat} passHref>
-    <ChatIcon />
-  </Link>
-);
-
-const FixedIconContainer = styled.div`
-  position: fixed;
-  z-index: 100;
-  bottom: ${(props) => props.theme.spacing[2]};
-  right: ${(props) => props.theme.spacing[2]};
-  
-  ${mediaQuery.minWidth.md} {
-    bottom: ${(props) => props.theme.spacing[4]};
-    right: ${(props) => props.theme.spacing[8]};
-  }
-`;
 
 export const FloatingChatButton = () => {
   const pathname = usePathname();
@@ -33,10 +13,10 @@ export const FloatingChatButton = () => {
   }
 
   return (
-    <FixedIconContainer>
+    <div className="fixed bottom-3 right-3 md:bottom-5 md:right-9 z-50">
       <Link aria-label="chat" target="_blank" href={slugs.chat} passHref >
         <ChatIcon />
       </Link>
-    </FixedIconContainer>
+    </div>
   );
 };
